[relate] Divide relate() and relation() code into interface and implementation.
Separate the optional code, place it in the details.
