+++ include/boost/geometry/algorithms/equals.hpp
@@ -51,7 +51,8 @@
 #include <boost/geometry/util/select_most_precise.hpp>
 
 #include <boost/geometry/algorithms/detail/equals/collect_vectors.hpp>
-#include <boost/geometry/algorithms/detail/relate/relate.hpp>
+#include <boost/geometry/algorithms/relate.hpp>
+#include <boost/geometry/algorithms/detail/relate/relate_base.hpp>
 
 #include <boost/geometry/views/detail/indexed_point_view.hpp>
