+++ include/boost/geometry/strategies/agnostic/relate.hpp
@@ -11,7 +11,7 @@
 #ifndef BOOST_GEOMETRY_STRATEGY_AGNOSTIC_RELATE_HPP
 #define BOOST_GEOMETRY_STRATEGY_AGNOSTIC_RELATE_HPP
 
-#include <boost/geometry/algorithms/detail/relate/relate.hpp>
+#include <boost/geometry/algorithms/relate.hpp>
 
 
 namespace boost { namespace geometry
